Extract composer markdown asset manager
