+++ assets/js/seo.js
@@ -1,8 +1,8 @@
 // seo.js - Dynamic SEO meta tag management for client-side routing
 // This maintains SEO benefits while keeping the "no compilation needed" philosophy
 
-import { getCurrentLang, DEFAULT_LANG } from './i18n.js?v=press-system-v3.4.28';
-import { getAvailableLangs } from './i18n.js?v=press-system-v3.4.28';
+import { getCurrentLang, DEFAULT_LANG } from './i18n.js?v=press-system-v3.4.29';
+import { getAvailableLangs } from './i18n.js?v=press-system-v3.4.29';
 import { parseFrontMatter } from './content.js';
 
 function ensureTrailingSlash(value) {
