+++ assets/press-system.json
@@ -1,13 +1,13 @@
 {
   "schemaVersion": 1,
   "type": "press-system",
-  "version": "3.4.28",
-  "tag": "v3.4.28",
+  "version": "3.4.29",
+  "tag": "v3.4.29",
   "upgradeFrom": {
     "ranges": [
-      ">=3.4.27 <3.4.28"
+      ">=3.4.28 <3.4.29"
     ],
     "allowUnknownSource": false,
-    "message": "Update to v3.4.27 first."
+    "message": "Update to v3.4.28 first."
   }
 }
