+++ scripts/test-system-release-package.sh
@@ -96,6 +96,11 @@ if ! grep -qx "press-system-${version}/assets/js/composer-site-settings-ui.js" "
   exit 1
 fi
 
+if ! grep -qx "press-system-${version}/assets/js/composer-markdown-assets.js" "${entries_file}"; then
+  echo "expected package to include composer Markdown asset manager code" >&2
+  exit 1
+fi
+
 if ! grep -qx "press-system-${version}/assets/js/editor-file-tree-ui.js" "${entries_file}"; then
   echo "expected package to include editor file tree UI code" >&2
   exit 1
